Improve handling of cached stores.
This includes major changes to how cached stores are managed.  To wit, the intention is
that all product stores are flushed from memory as soon as they are no longer needed.
The current algorithm is that all stores cached by the node are evicted whenever:

 (a) any non-reduction node has both processed a particular level and received a flush
     store for that level, or
 (b) any reduction node has received a flush store for a given level and processed all the
     children stores indicated by the flush store ID.

This algorithm is subject to early eviction of cached stores if downstream nodes have not
completed their processing of a cached store whenever a flush is emitted for that cached
store.  When that happens, the store is regenerated, resulting in more calls to the user
function than are expected.  This should be fixed at some point.

This commit also includes:

 - Removal of unnecessary 'setup' stage.
 - More user-friendly interface for sources: the user no longer has to supply the flush
   stores.  The downside is that the implementation is slower.  Added a couple FIXME
   comments to describe the situation.